+++ src/extdefs/extdefs.jl
@@ -2,6 +2,7 @@
 
 include("advancedhmc_defs.jl")
 include("cuba_defs.jl")
+include("mgvi_defs.jl")
 include("nestedsamplers_defs.jl")
 include("optim_defs.jl")
 include("optimization_defs.jl")