+++ test/samplers/test_samplers.jl
@@ -7,4 +7,5 @@ Test.@testset "samplers" begin
     include("mcmc/test_mcmc.jl")
     include("importance/test_importance_sampler.jl")
     include("nested_sampling/test_nested_sampling.jl")
+    include("test_mgvi.jl")
 end